services/api/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
uvloop==0.19.0  # High-performance event loop
httptools==0.6.1  # Fast HTTP parser

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Firebase & GCP
firebase-admin==6.3.0
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.17.0

# Redis (supports Upstash with TLS)
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Message Queue (optional)
aio-pika==9.4.0

# Logging
structlog==24.1.0

# Production utilities
orjson==3.9.12  # Fast JSON serialization
